Ignore broken pipe errors in calls to splice()
In some cases the receiver can close its end of the pipe early (e.g.,
if it independently determines it has reached the end of the stream),
before the flow pump can confirm there is no more data to process.
In this case splice would raise a BrokenPipeError, which is now safely
ignored.
This does not apply to the first call, when it is correct to raise an
error.